requirements.txt
streamlit
faiss-cpu
numpy
PyPDF2
sentence-transformers
groq